Use WTF::Pointer for pointer bitwise operations
in the following components/classes of JavaScriptCore:
 - bytecode/LLIntCallLinkInfo
 - bytecode/PropertyCondition
 - heap/CellContainer
 - heap/LargeAllocation
 - runtime/InferredValue
 - runtime/JSCJSValue
 - runtime/JSString
 - runtime/LazyProperty
 - runtime/RegExpObject
 - runtime/StructureTransitionTable
 - runtime/SymbolTable
 - wasm/WasmIndexOrName

+++ Source/JavaScriptCore/runtime/StructureTransitionTable.h
@@ -142,7 +142,7 @@ inline bool setsReadOnlyOnNonAccessorProperties(NonPropertyTransition transition
 }
 
 class StructureTransitionTable {
-    static constexpr intptr_t UsingSingleSlotFlag = 1;
+    static constexpr unsigned UsingSingleSlotFlag = 1;
 
     
     struct Hash {
